+++ commands/editing/R.cpp
@@ -10,7 +10,7 @@ void VM::R::executeFor(VM::VMModel &model, int times) {
     std::pair<int, int> cursor = state->getCursor();
     std::string line = state->getLine(cursor.first);
     char c = model.getInput().back();
-    if (cursor.second + times <= line.length()) {
+    if (cursor.second + times <= (int) line.length()) {
         std::string newLine = line.substr(0, cursor.second);
         int i = cursor.second;
         for (; i < cursor.second + times; ++i) {
