+++ commands/ex/ExR.cpp
@@ -9,7 +9,7 @@ VM::ExR::ExR() {
 void VM::ExR::executeFor(VM::VMModel &model, const std::string &inputSequence) {
     WindowState *state = model.getActiveState();
     std::string newFileName;
-    for (int i = 2; i < inputSequence.length(); ++i) {
+    for (size_t i = 2; i < inputSequence.length(); ++i) {
         if (inputSequence.at(i) != ' ') {
             newFileName = inputSequence.substr(i);
             break;
